Fixed incorrect reading of the "mount" command output in the analysis function of mounted file systems, eliminated writing unnecessary data to the buffer. Also fixed a possible blocking of a child process due to an overflow of the native I/O stream and, as a result, exit from waiting by timeout without receiving any data: https://developer.android.com/reference/java/lang/Process
